facebook_automation/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
crewai[tools,google-genai]==1.7.2
beautifulsoup4>=4.12.0
requests>=2.31.0
litellm>=1.75.3
pydantic>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pytz>=2024.1
pyyaml>=6.0.0
google-generativeai>=0.3.0
